fold in getargs
