+++ packages/ui-components/src/lib/components/deployment/DeploymentSteps.svelte
@@ -28,7 +28,8 @@
 	import DeploymentSectionHeader from './DeploymentSectionHeader.svelte';
 	import { useGui } from '$lib/hooks/useGui';
 	import { useAccount } from '$lib/providers/wallet/useAccount';
-	import { handleDeployment, type DeploymentHandlers } from '../../utils/handleDeployment';
+	import { handleDeployment } from '../../utils/handleDeployment';
+	import type { HandleAddOrderResult } from './getDeploymentTransactionArgs';
 
 	interface Deployment {
 		key: string;
@@ -41,10 +42,11 @@
 	/** Strategy details containing name and description configuration */
 	export let strategyDetail: NameAndDescriptionCfg;
 	/** Handlers for deployment modals */
-	export let deploymentHandlers: DeploymentHandlers;
+	export let handleClickDeploy: (handleAddOrderResult: HandleAddOrderResult) => void;
 	export let wagmiConnected: Writable<boolean>;
 	export let appKitModal: Writable<AppKit>;
 	export let settings: Writable<ConfigSource>;
+	export let registryUrl: string;
 
 	let allDepositFields: GuiDepositCfg[] = [];
 	let allTokenOutputs: OrderIOCfg[] = [];
@@ -218,7 +220,7 @@
 				return;
 			}
 			DeploymentStepsError.clear();
-			await handleDeployment(gui, $account, deploymentHandlers, subgraphUrl);
+			await handleDeployment(gui, $account, subgraphUrl);
 		} catch (e) {
 			DeploymentStepsError.catch(e, DeploymentStepsErrorCode.ADD_ORDER_FAILED);
 		} finally {
@@ -227,12 +229,20 @@
 	}
 </script>
 
-<div class="py-6">
+<div>
+	{#if $deploymentStepsError}
+		<Alert color="red">
+			<p class="text-red-500">{$deploymentStepsError.code}</p>
+			{#if $deploymentStepsError.details}
+				<p class="text-red-500">{$deploymentStepsError.details}</p>
+			{/if}
+		</Alert>
+	{/if}
 	{#if gui}
 		<div class="flex max-w-3xl flex-col gap-12" in:fade>
 			{#if deployment}
 				<div class="flex max-w-2xl flex-col gap-4 text-start">
-					<h1 class=" text-4xl font-semibold text-gray-900 lg:text-6xl dark:text-white">
+					<h1 class="text-4xl font-semibold text-gray-900 lg:text-6xl dark:text-white">
 						{strategyDetail.name}
 					</h1>
 					<p class="text-xl text-gray-600 lg:text-2xl dark:text-gray-400">
@@ -241,57 +251,77 @@
 				</div>
 			{/if}
 
-			<Toggle bind:checked={showAdvancedOptions}>Show advanced options</Toggle>
-
-			{#if showAdvancedOptions}
-				{#each allFieldDefinitionsWithDefaults as fieldDefinition}
-					<FieldDefinitionInput {fieldDefinition} />
-				{/each}
-
-				{#each allDepositFields as deposit}
-					<DepositInput {deposit} {gui} />
-				{/each}
-
-				{#each allTokenInputs as input, i}
-					<TokenIOInput {i} label="Input" vault={input} {gui} />
-				{/each}
-
-				{#each allTokenOutputs as output, i}
-					<TokenIOInput {i} label="Output" vault={output} {gui} />
-				{/each}
+			{#if selectTokens && selectTokens.length > 0}
+				<div class="flex w-full flex-col gap-4">
+					<DeploymentSectionHeader
+						title="Select Tokens"
+						description="Select the tokens that you want to use in your order."
+					/>
+					{#each selectTokens as token}
+						<SelectToken {token} {onSelectTokenSelect} />
+					{/each}
+				</div>
 			{/if}
 
-			{#if $deploymentStepsError}
-				<Alert color="red">
-					<p class="text-red-500">{$deploymentStepsError.code}</p>
-					{#if $deploymentStepsError.details}
-						<p class="text-red-500">{$deploymentStepsError.details}</p>
-					{/if}
-				</Alert>
-			{/if}
-
-			<div class="flex flex-wrap items-start justify-start gap-2">
-				{#if $account}
-					<Button
-						data-testid="deploy-button"
-						size="lg"
-						disabled={checkingDeployment}
-						on:click={handleDeployButtonClick}
-						class="bg-gradient-to-br from-blue-600 to-violet-600"
-					>
-						{#if checkingDeployment}
-							<Spinner size="4" color="white" />
-							<span class="ml-2">Checking deployment...</span>
-						{:else}
-							Deploy Strategy
-						{/if}
-					</Button>
-				{:else}
-					<WalletConnect {appKitModal} connected={wagmiConnected} />
+			{#if allTokensSelected || selectTokens?.length === 0}
+				{#if allFieldDefinitionsWithoutDefaults.length > 0}
+					{#each allFieldDefinitionsWithoutDefaults as fieldDefinition}
+						<FieldDefinitionInput {fieldDefinition} />
+					{/each}
 				{/if}
-				<ComposedRainlangModal {gui} />
-				<ShareChoicesButton handleShareChoices={_handleShareChoices} />
-			</div>
+
+				<Toggle bind:checked={showAdvancedOptions}>Show advanced options</Toggle>
+
+				{#if showAdvancedOptions}
+					{#each allFieldDefinitionsWithDefaults as fieldDefinition}
+						<FieldDefinitionInput {fieldDefinition} />
+					{/each}
+
+					{#each allDepositFields as deposit}
+						<DepositInput {deposit} />
+					{/each}
+
+					{#each allTokenInputs as input, i}
+						<TokenIOInput {i} label="Input" vault={input} />
+					{/each}
+
+					{#each allTokenOutputs as output, i}
+						<TokenIOInput {i} label="Output" vault={output} />
+					{/each}
+				{/if}
+
+				{#if $deploymentStepsError}
+					<Alert color="red">
+						<p class="text-red-500">{$deploymentStepsError.code}</p>
+						{#if $deploymentStepsError.details}
+							<p class="text-red-500">{$deploymentStepsError.details}</p>
+						{/if}
+					</Alert>
+				{/if}
+
+				<div class="flex flex-wrap items-start justify-start gap-2">
+					{#if $account}
+						<Button
+							data-testid="deploy-button"
+							size="lg"
+							disabled={checkingDeployment}
+							on:click={handleDeployButtonClick}
+							class="bg-gradient-to-br from-blue-600 to-violet-600"
+						>
+							{#if checkingDeployment}
+								<Spinner size="4" color="white" />
+								<span class="ml-2">Checking deployment...</span>
+							{:else}
+								Deploy Strategy
+							{/if}
+						</Button>
+					{:else}
+						<WalletConnect {appKitModal} connected={wagmiConnected} />
+					{/if}
+					<ComposedRainlangModal />
+					<ShareChoicesButton handleShareChoices={_handleShareChoices} />
+				</div>
+			{/if}
 		</div>
 	{/if}
 </div>
